Initial refactoring of the code. Enable the ability to set the state of the widget.
